pyproject.toml
[tool.poetry]
name = "agv-prototype"
version = "0.1.0"
description = ""
authors = ["drmwnrafi <drmwnnrafi@gmail.com>"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pyserial = "^3.5"
crcmod = "^1.7"
pymodbus = "^3.5.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
